pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "persian-jalali-calendar"
dynamic = ["version"]
authors = [ { name="Hamid R. Yari", email="hryarih@gmail.com" } ]
description = "A simple and accurate Python library for the Persian Jalali calendar."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
dependencies = [
    "backports.zoneinfo; python_version < '3.9'",
    "tzdata"
]
keywords = [
    "jalali", "persian", "calendar", "shamsi", "khorshidi",
    "date", "converter", "datetime", "timezone"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "jalali_calendar.__version__" }

[project.urls]
"Homepage" = "https://github.com/hryarih32/jalali-calendar"
"Bug Tracker" = "https://github.com/hryarih32/jalali-calendar/issues"